Simplified serverchannelbase, hopefully fixing intermittent bugs.
The FTP server needed a way to queue a task in the channel.  It used 
start_task, but start_task had no way to queue tasks correctly if tasks 
already happened to be running.  So an assertion failure resulted 
occasionally.

Now, there is a queue of tasks rather than a queue of requests.  
Anything that needs to can send a task to a channel.  The task will be 
executed in synchronous mode.

Also, the basic request parsing is now done in the asyncore main thread.  
When SimultaneousModeChannel was dropped, it became impossible to receive 
data in application threads anyway.