Implement OverflowableBuffer in terms of SpooledTemporaryFile
Fixes #5

Even for small buffers (the 8192 that previously were guaranteed to be
kept in a simple string) this should be faster, depending on the
number of writes possibly. Previously the strbuf was built up with
concatenation: `buf = buf + data`. Now we go directly into BytesIO and
avoid the extra copying.

Microbenchmarks on Python 3.6:

In [1]: def append():
   ...:     buf = b''
   ...:     while len(buf) < 8192:
   ...:         buf += b'data' * 100
   ...:

In [4]: %timeit append()
10.9 µs ± 268 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [8]: def bio():
   ...:     b = io.BytesIO()
   ...:     cnt = 0
   ...:     while cnt < 8192:
   ...:         cnt += b.write(b'data'*100)
   ...:

In [9]: bio()

In [10]: %timeit bio()
8.55 µs ± 282 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)